iftpfm2

Module logging

Statics

LOG_FILE

Global log file path protected by Mutex

Functions

log

Logs a message to either a file or stdout

log_with_thread

Logs a message with timestamp and optional thread ID

set_log_file

Sets the path for the log file