iftpfm2::logging

Function log

pub fn log(message: &str) -> Result<()>

Logs a message to either a file or stdout

This function takes a message as input and logs it with a timestamp. If a log file has been set (using set_log_file), the message is appended to that file. Otherwise, the message is printed to stdout.

Arguments

message - The message to be logged

Returns

io::Result<()> - Ok if the logging was successful, Err otherwise Logs a message.

Logs a message with timestamp to configured output

Arguments

message - The message to log

Returns

io::Result<()> - Ok on success, Err if writing fails

Example

// log("Starting transfer").unwrap();