iftpfm2::protocols

Enum Client

pub enum Client {
    Ftp(FtpClient),
    Ftps(FtpsClient),
    Sftp(SftpClient),
}

Enum wrapper for FTP/FTPS/SFTP clients

Since FileTransferClient has generic methods, it cannot be used as dyn FileTransferClient. This enum provides a concrete type that can be used to hold FTP, FTPS, or SFTP clients.

Variants

Ftp(FtpClient)

Ftps(FtpsClient)

Sftp(SftpClient)

Implementations

impl Client

pub fn connect( proto: &Protocol, host: &str, port: u16, timeout: Duration, insecure_skip_verify: bool, user: &str, password: Option<&str>, keyfile_path: Option<&str>, ) -> Result<Self, FtpError>

Connect to a server and create a client of the appropriate type

pub fn login(&mut self, user: &str, password: &str) -> Result<(), FtpError>

Authenticate with username and password

pub fn cwd(&mut self, path: &str) -> Result<(), FtpError>

Change working directory

pub fn transfer_type(&mut self, mode: TransferMode) -> Result<(), FtpError>

Set transfer type (binary or ASCII)

pub fn nlst(&mut self, path: Option<&str>) -> Result<Vec<String>, FtpError>

Get list of files in directory (NLST command)

pub fn mdtm(&mut self, filename: &str) -> Result<NaiveDateTime, FtpError>

Get file modification time (MDTM command)

pub fn size(&mut self, filename: &str) -> Result<usize, FtpError>

Get file size (SIZE command)

pub fn retr<F, D>(&mut self, filename: &str, callback: F) -> Result<D, FtpError>
where F: FnMut(&mut dyn Read) -> Result<D, FtpError>,

Retrieve file contents

pub fn put_file<R: Read>( &mut self, filename: &str, reader: &mut R, ) -> Result<u64, FtpError>

Upload file contents

pub fn rename(&mut self, from: &str, to: &str) -> Result<(), FtpError>

Rename a file

pub fn rm(&mut self, filename: &str) -> Result<(), FtpError>

Remove/delete a file

pub fn quit(self) -> Result<(), FtpError>

Quit/disconnect from the server